IPV6 peer creation fix

+++ src/test/java/com/wireguard/external/wireguard/SubnetTest.java
@@ -3,6 +3,8 @@ package com.wireguard.external.wireguard;
 import com.wireguard.external.network.Subnet;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigInteger;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -24,7 +26,7 @@ class SubnetTest {
         Subnet subnet = Subnet.valueOf("192.168.0.3/24");
         assertEquals("192.168.0.255", subnet.getLastIpString(), ".getLastIpString()");
         assertEquals("192.168.0.0", subnet.getFirstIpString(), ".getFirstIpString()");
-        assertEquals(256, subnet.getIpCount(), ".getIpCount()");
+        assertEquals(BigInteger.valueOf(256), subnet.getIpCount(), ".getIpCount()");
         assertEquals(24, subnet.getNumericMask(), ".getNumericMask()");
         assertEquals(3232235775L, subnet.getLastIpNumeric(), ".getLastIpNumeric()");
         assertEquals(3232235520L, subnet.getFirstIpNumeric(), ".getFirstIpNumeric()");
@@ -35,7 +37,7 @@ class SubnetTest {
         Subnet subnet = Subnet.valueOf("0.0.0.0/0");
         assertEquals("0.0.0.0", subnet.getFirstIpString(), ".getFirstIpString()");
         assertEquals("255.255.255.255", subnet.getLastIpString(), ".getLastIpString()");
-        assertEquals(4294967296L, subnet.getIpCount(), ".getIpCount()");
+        assertEquals(BigInteger.valueOf(4294967296L), subnet.getIpCount(), ".getIpCount()");
         assertEquals(0, subnet.getNumericMask(), ".getNumericMask()");
         assertEquals(0L, subnet.getFirstIpNumeric(), ".getFirstIpNumeric()");
         assertEquals(4294967295L, subnet.getLastIpNumeric(), ".getLastIpNumeric()");
